wsh_shell_adapter/requirements.txt
pyserial==3.5
toml==0.10.2
